Add the Drop trait to a generated hub to prevent unsoundness
A generated hub could have its constituents extracted via

        let Hub { x, y, .. } = hub;

Since topics are Send even though they contain an Rc, this could cause
UB if x was sent to one thread and y to another: as dropping them can
race due to the dependency on the manager.

An alternative was

        let x = hub.x;

But this also no longer works with Drop implemented.