Mitigate decoy-based mutable aliasing attacks
By constructing a type with a field that overlaps in the pointer value
of its parent we are able to mutably borrow an object twice because
suspend doesn't detect that we're suspending a field, and not the proper
object that was initially emitted:

        struct Decoy {
                a: (),
                b: u8,
        }

        let my_node = Node::new(Decoy { a: (), b: 0 });

        my_node.emit(|x| {
                // Works because `x.a` and `x` _may_ share the same address.
                let x_b = &mut x.b;
                x.a.suspend(|| {
                        my_node.emit(|y| {
                                let y_b = &mut y.b

                                // x_b and y_b are mutably aliasing!
                        });
                });
        });

This patch adds the size value to the stack, because we simply can not
overlap on dynamic allocations, nor can a field be bigger than the class
itself. This ensures we use the correct object.